chore: refactor sidebar components
