+++ src/views/search/components/filter-section.module.scss
@@ -1,0 +1,156 @@
+@import "src/tokens/index";
+
+:root {
+	--search-page_filter_sidebar_padding-end: 0px;
+
+	@include from($tabletSmall) {
+		--search-page_filter_sidebar_max-width: 288px;
+	}
+
+	@include from($desktopSmall) {
+		--search-page_filter_sidebar_max-width: 320px;
+		--search-page_filter_sidebar_padding-top: var(--spc-4x);
+		--search-page_filter_sidebar_padding-bottom: var(--site-spacing);
+		--search-page_filter_sidebar_padding-start: var(--site-spacing);
+		--search-page_filter_sidebar_gap: var(--spc-4x);
+		--search-page_filter_sidebar_order-btn-gap: var(--spc-2x);
+
+		--search-page_filter_list_header_min-height: var(--min-target-size_l);
+		--search-page_filter_list_header_gap: var(--spc-2x);
+		--search-page_filter_list_header_padding-start: var(--spc-3x);
+		--search-page_filter_list_header_padding-end: var(--spc-2x);
+
+		--search-page_filter_list_header_title_color: var(
+			--foreground_emphasis-high
+		);
+		--search-page_filter_list_header_count_color: var(--primary_default);
+		--search-page_filter_list_header_icon_color: var(--primary_default);
+		--search-page_filter_list_header_icon_size: var(--icon-size_regular);
+
+		--search-page_filter_list_divider_color: var(--background_disabled);
+		--search-page_filter_list_divider_width: var(--border-width_s);
+
+		--search-page_filter_list_more-btn_margin-horizontal: var(--spc-4x);
+		--search-page_filter_list_more-btn_margin-vertical: var(--spc-4x);
+	}
+
+	@include from($desktopLarge) {
+		--search-page_filter_sidebar_max-width: 360px;
+		--search-page_filter_sidebar_padding-end: var(--site-spacing);
+	}
+}
+
+:root {
+	--search-page_filter_list_item_padding-vertical: var(--spc-1x);
+	--search-page_filter_list_item_padding-start: var(--spc-3x);
+	--search-page_filter_list_item_padding-end: var(--spc-2x);
+	--search-page_filter_list_item_icon-size: var(--icon-size_regular);
+	--search-page_filter_list_item_gap: var(--spc-2x);
+
+	--search-page_filter_list_item_label_color: var(--foreground_emphasis-high);
+	--search-page_filter_list_item_count_color: var(--foreground_disabled);
+
+	--search-page_filter_list_item_background-color_hovered: var(
+		--surface_emphasis-low
+	);
+	--search-page_filter_list_item_background-color_pressed: var(
+		--surface_emphasis-medium
+	);
+	--search-page_filter_list_item_background-color_selected: var(
+		--surface_emphasis-none
+	);
+}
+
+.section {
+	display: flex;
+	flex-direction: column;
+	border: var(--search-page_filter_list_divider_width) solid
+		var(--search-page_filter_list_divider_color);
+	border-radius: var(--corner-radius_l);
+	overflow: hidden;
+
+	flex-grow: 0;
+	@include transition(flex-grow);
+
+	&Expanded {
+		flex-grow: 1;
+	}
+}
+
+.sectionHeader {
+	display: grid;
+	grid-template-columns: 1fr auto;
+	grid-template-rows: 1fr;
+}
+
+.clearContainer {
+	grid-column: 2;
+	grid-row: 1;
+	display: flex;
+	align-items: center;
+	justify-content: flex-end;
+}
+
+.clearChip {
+	border-radius: var(--corner-radius_circular);
+	margin-right: var(--search-page_filter_list_header_padding-end);
+	margin-left: var(--search-page_filter_list_header_gap);
+}
+
+.sectionTitle {
+	grid-column: 1 / span 2;
+	grid-row: 1;
+
+	display: flex;
+	gap: var(--search-page_filter_list_header_gap);
+	width: 100%;
+	// Replace with tokens
+	padding-top: var(--spc-2x);
+	padding-bottom: var(--spc-2x);
+	padding-left: var(--search-page_filter_list_header_padding-start);
+	padding-right: var(--search-page_filter_list_header_padding-end);
+
+	// Stop being a button 😡
+	background: unset;
+	border: unset;
+	outline: unset;
+	font-weight: unset;
+	font-size: unset;
+	line-height: unset;
+}
+
+.collapseIcon {
+	background: var(--search-page_filter_list_header_icon_color);
+	height: var(--search-page_filter_list_header_icon_size);
+	width: var(--search-page_filter_list_header_icon_size);
+	mask: url('data:image/svg+xml,<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M5.29289 9.29289C5.68342 8.90237 6.31658 8.90237 6.70711 9.29289L12 14.5858L17.2929 9.29289C17.6834 8.90237 18.3166 8.90237 18.7071 9.29289C19.0976 9.68342 19.0976 10.3166 18.7071 10.7071L13.4142 16C12.6332 16.781 11.3668 16.781 10.5858 16L5.29289 10.7071C4.90237 10.3166 4.90237 9.68342 5.29289 9.29289Z" fill="%23006590"/></svg>');
+	mask-position: center;
+	mask-repeat: no-repeat;
+	transform: rotate(180deg);
+	transition: transform 0.2s ease-in-out;
+}
+
+.collapsed {
+	transform: rotate(0deg);
+}
+
+.sectionTitleText {
+	color: var(--search-page_filter_list_header_title_color);
+}
+
+.sectionNumberText {
+	color: var(--search-page_filter_list_header_count_color);
+}
+
+.sectionContent {
+	overflow: auto;
+	height: 1px;
+	flex-grow: 1;
+
+	opacity: 1;
+	@include transition(opacity);
+
+	&[aria-hidden="true"] {
+		opacity: 0;
+	}
+}
